The AGRI-Book Publishing is a web portal for an agriculture publication house. The system is used to fulfil the whole publishing process and bookseller can contact with the publisher. The main objective for making this project is for the publisher to take control over the publishing procedure. The publisher has full access over the users of the system. The users over which the publisher has full access are	Author, Editor, Designer and Bookseller.
The AGRI-Book Publishing is a web portal for an agriculture publication house. The system is used to fulfil the whole publishing process and bookseller can contact with the publisher.
The main objective for making this project is for the publisher to take control over the publishing procedure. The publisher has full access over the users of the system. The users over which the publisher has full access are	Author, Editor, Designer and Bookseller.